Ignoring pager HTML validation error
